Get file content length
We need to allow the submitter to know the size of attachments before
downloading them.

This stems from previous work which grouped attachments by file size
into the lowest number of emails as possible.

We now need the submitter to be able to know how to group the
attachments before downloading them so we can prevent any failed delayed
jobs from resending any emails that have been successfully sent to a
recipient.