Update dependencies and increase code coverage
While working on the filestore timeouts I notice some deprecation warnings, as well as some old dependencies.

I've bumped ruby/rails and others, and added simplecov to track coverage. I've set it to 98 after increasing it a bit with a few additional tests.

Tweaked a bit the upload logging debug.

Also, disabled the wrapping of parameters as this is not necessary and was causing doubling the payload submitted to sentry.