Rename errors without a concrete validator
Any errors returned by the uploads controller need to be "understood" by the different validators implemented in the metadata presenter.

If the validator is not trigger for a given error, it is considered "valid" and has the consequence it produces an unhandled error on the runner/frontend instead of a "nice" error like it should.

These are rare errors majority of the time will not happen, but if any of them happen now they will show a generic "nice error" (requires new version of metadata presenter in the runner, separate PRs).